TestTestgresCommon is updated
New tests are added:
 - test_auto_name
 - test_file_tail
 - test_isolation_levels
