+++ tests/test_testgres_common.py
@@ -5,7 +5,9 @@ from .helpers.os_ops_descrs import OsOperations
 from ..testgres.node import PgVer
 from ..testgres.node import PostgresNode
 from ..testgres.utils import get_pg_version2
+from ..testgres.utils import file_tail
 from ..testgres import ProcessType
+from ..testgres import IsolationLevel
 from ..testgres import TestgresException
 from ..testgres import StartNodeException
 from ..testgres import QueryException
@@ -15,6 +17,7 @@ import pytest
 import six
 import logging
 import time
+import tempfile
 
 
 class TestTestgresCommon:
@@ -185,6 +188,63 @@ class TestTestgresCommon:
         str(StartNodeException('msg', [('file', 'lines')]))
         str(ExecUtilException('msg', 'cmd', 1, 'out'))
         str(QueryException('msg', 'query'))
+
+    def test_auto_name(self, os_ops: OsOperations):
+        assert isinstance(os_ops, OsOperations)
+
+        with __class__.helper__get_node(os_ops).init(allow_streaming=True).start() as m:
+            with m.replicate().start() as r:
+                # check that nodes are running
+                assert (m.status())
+                assert (r.status())
+
+                # check their names
+                assert (m.name != r.name)
+                assert ('testgres' in m.name)
+                assert ('testgres' in r.name)
+
+    def test_file_tail(self):
+        s1 = "the quick brown fox jumped over that lazy dog\n"
+        s2 = "abc\n"
+        s3 = "def\n"
+
+        with tempfile.NamedTemporaryFile(mode='r+', delete=True) as f:
+            sz = 0
+            while sz < 3 * 8192:
+                sz += len(s1)
+                f.write(s1)
+            f.write(s2)
+            f.write(s3)
+
+            f.seek(0)
+            lines = file_tail(f, 3)
+            assert (lines[0] == s1)
+            assert (lines[1] == s2)
+            assert (lines[2] == s3)
+
+            f.seek(0)
+            lines = file_tail(f, 1)
+            assert (lines[0] == s3)
+
+    def test_isolation_levels(self, os_ops: OsOperations):
+        assert isinstance(os_ops, OsOperations)
+        with __class__.helper__get_node(os_ops).init().start() as node:
+            with node.connect() as con:
+                # string levels
+                con.begin('Read Uncommitted').commit()
+                con.begin('Read Committed').commit()
+                con.begin('Repeatable Read').commit()
+                con.begin('Serializable').commit()
+
+                # enum levels
+                con.begin(IsolationLevel.ReadUncommitted).commit()
+                con.begin(IsolationLevel.ReadCommitted).commit()
+                con.begin(IsolationLevel.RepeatableRead).commit()
+                con.begin(IsolationLevel.Serializable).commit()
+
+                # check wrong level
+                with pytest.raises(expected_exception=QueryException):
+                    con.begin('Garbage').commit()
 
     @staticmethod
     def helper__get_node(os_ops: OsOperations, name=None):
